requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0

# API integrations
websockets>=11.0.0
aiohttp>=3.8.0

# AI/LLM providers
google-generativeai>=0.3.0
openai>=1.0.0

# News data sources (using requests for CryptoPanic)
newsapi-python>=0.2.6

# Utilities
pytz>=2023.3
python-dateutil>=2.8.0
retrying>=1.3.4

# Development & testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
